feat: add news page component and route
Refactor news page into separate component and update route to use it
